+++ src/Fragments/Tests/App/ContextActions/SyncContextSitesTest.php
@@ -12,12 +12,19 @@ class SyncContextSitesTest extends ChiefTestCase
     public function test_it_can_sync_context_sites()
     {
         $owner = $this->setUpAndCreateArticle();
-        $context = FragmentTestHelpers::createContext($owner, ['default', 'site-b'], ['default', 'site-b']);
+        $context = FragmentTestHelpers::createContext($owner, ['site-a', 'site-b'], ['site-b']);
+        $context2 = FragmentTestHelpers::createContext($owner, ['site-a', 'site-b'], ['site-a']);
 
         $this->assertDatabaseHas('contexts', [
             'id' => $context->id,
-            'allowed_sites' => json_encode(['default', 'site-b']),
-            'active_sites' => json_encode(['default', 'site-b']),
+            'allowed_sites' => json_encode(['site-a', 'site-b']),
+            'active_sites' => json_encode(['site-b']),
+        ]);
+
+        $this->assertDatabaseHas('contexts', [
+            'id' => $context2->id,
+            'allowed_sites' => json_encode(['site-a', 'site-b']),
+            'active_sites' => json_encode(['site-a']),
         ]);
 
         app(ContextApplication::class)->syncSites(
@@ -29,12 +36,19 @@ class SyncContextSitesTest extends ChiefTestCase
             'allowed_sites' => json_encode(['site-b']),
             'active_sites' => json_encode(['site-b']),
         ]);
+
+        $this->assertDatabaseHas('contexts', [
+            'id' => $context2->id,
+            'allowed_sites' => json_encode(['site-b']),
+            'active_sites' => json_encode([]),
+        ]);
     }
 
     public function test_it_does_remove_active_site_if_site_no_longer_allowed()
     {
         $owner = $this->setUpAndCreateArticle();
-        $context = FragmentTestHelpers::createContext($owner, ['default', 'site-b'], ['default']);
+        $context = FragmentTestHelpers::createContext($owner, ['site-a', 'site-b'], ['site-a']);
+        $context2 = FragmentTestHelpers::createContext($owner, ['site-a', 'site-b'], ['site-a']);
 
         app(ContextApplication::class)->syncSites(
             new SyncSites($owner->modelReference(), ['site-b'])
@@ -45,21 +59,50 @@ class SyncContextSitesTest extends ChiefTestCase
             'allowed_sites' => json_encode(['site-b']),
             'active_sites' => json_encode([]),
         ]);
+
+        $this->assertDatabaseHas('contexts', [
+            'id' => $context2->id,
+            'allowed_sites' => json_encode(['site-b']),
+            'active_sites' => json_encode([]),
+        ]);
     }
 
-    public function test_it_does_not_add_new_allowed_sites_to_existing_contexts()
+    public function test_it_syncs_allowed_sites_when_only_one_context()
     {
         $owner = $this->setUpAndCreateArticle();
-        $context = FragmentTestHelpers::createContext($owner, ['default'], ['default']);
+        $context = FragmentTestHelpers::createContext($owner, ['site-a'], ['site-a']);
 
         app(ContextApplication::class)->syncSites(
-            new SyncSites($owner->modelReference(), ['default', 'site-b'])
+            new SyncSites($owner->modelReference(), ['site-a', 'site-b'])
         );
 
         $this->assertDatabaseHas('contexts', [
             'id' => $context->id,
-            'allowed_sites' => json_encode(['default']),
-            'active_sites' => json_encode(['default']),
+            'allowed_sites' => json_encode(['site-a', 'site-b']),
+            'active_sites' => json_encode(['site-a', 'site-b']),
+        ]);
+    }
+
+    public function test_it_does_not_add_new_allowed_sites_to_contexts_when_more_than_one_context()
+    {
+        $owner = $this->setUpAndCreateArticle();
+        $context = FragmentTestHelpers::createContext($owner, ['site-a'], ['site-a']);
+        $context2 = FragmentTestHelpers::createContext($owner, ['site-a', 'site-b'], ['site-b']);
+
+        app(ContextApplication::class)->syncSites(
+            new SyncSites($owner->modelReference(), ['site-a', 'site-b'])
+        );
+
+        $this->assertDatabaseHas('contexts', [
+            'id' => $context->id,
+            'allowed_sites' => json_encode(['site-a']),
+            'active_sites' => json_encode(['site-a']),
+        ]);
+
+        $this->assertDatabaseHas('contexts', [
+            'id' => $context2->id,
+            'allowed_sites' => json_encode(['site-a', 'site-b']),
+            'active_sites' => json_encode(['site-b']),
         ]);
     }
 }